feat(cc-pricing-page)!: update styles and behavior, remove smart
BREAKING CHANGE: We needed to be able to style and move the pricing
components inside the `cc-pricing-page` component a lot more.
For instance, the `cc-pricing-header` can now be positioned wherever one
wants within the `cc-pricing-page` because everything is slotted within
the `cc-pricing-page`.
The smart `cc-pricing-page` smart component was mainly interacting
with `cc-pricing-header` data (fetching `zones`).
This is why we removed the `cc-pricing-page.smart` part and added a
`cc-pricing-header.smart`.