chore(check-type-imports): remove unused imports checking feature
We need to be able to import types that are supposed to be used across
the whole file instead of the constructor only.
Unused types are checked by the `tsserver` within our IDEs so this part of
the task is obsolete.

chore(tasks): remove `check-type-imports`

This is now covered by typechecking enabled in our IDEs and progressively in the CI (see #916)

chore(check-type-imports): remove unused imports checking feature

We need to be able to import types that are supposed to be used across
the whole file instead of the constructor only.
Unused types are checked by the `tsserver` within our IDEs so this part of
the task is obsolete.

chore(tsconfig): feedback fix config (fixup later)

chore(typechecking): feedback restore the check type imports task (fixup later)