refactor(.storybook/.htaccess): remove useless RewriteRule directive
This RewriteRule doesn't work for the prod app because the storybook
assets are in `/doc/clever-components` and not directly at the root.

Ideally we would go for a `RewriteRule` that would take this into
account but we cannot access the app environment variables from the
`.htaccess` file so we would have to do something like this:

`RewriteRule ^ /doc/clever-components/index.html`

With this, any path after `doc/clever-components/` would be redirected
to the `index.html`.

But we don't really want to maintain this hardcoded value.

Since all storybook pages rely on a queryParam on the root URL anyway,
we don't really have to support this.